requirements.txt
mlflow==2.13.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
pytest==8.2.0
flake8==7.0.0
